refactor: change epochs logic to not change the initial deposit date but use the initial date and a changing deposit as coordinates
added a bunch of test, but still work in progress
other tests are failing and UserData logic needs to change

+++ src/subscription/Lib.sol
@@ -63,8 +63,8 @@ library Lib {
     }
 
 
-    function expiresAt(uint256 amount, uint256 depositedAt, uint256 mRate) internal pure returns (uint256) {
-        return depositedAt + (amount / mRate);
+    function expiresAt(uint256 amount, uint64 depositedAt, uint256 mRate) internal pure returns (uint64) {
+        return depositedAt + uint64(amount / mRate);
     }
 
     function multipliedRate(uint256 rate, uint256 multiplier) internal pure returns (uint256) {
